[Arch] split ArchAxis into separate modules
- remove unused import from resulting ArchAxisSystem and ArchGrid
- move inline imports to file import block
